+++ Sources/Sentry/include/SentrySdkInfo.h
@@ -1,5 +1,10 @@
-#import "SentrySerializable.h"
 #import <Foundation/Foundation.h>
+
+#if __has_include(<Sentry/SentrySerializable.h>)
+#    import <Sentry/SentrySerializable.h>
+#else
+#    import "SentrySerializable.h"
+#endif
 
 NS_ASSUME_NONNULL_BEGIN