+++ Sources/SentryCrash/Recording/SentryCrashReport.c
@@ -33,7 +33,6 @@
 #include "SentryCrashJSONCodec.h"
 #include "SentryCrashMach.h"
 #include "SentryCrashMemory.h"
-#include "SentryCrashMonitor_Zombie.h"
 #include "SentryCrashObjC.h"
 #include "SentryCrashReportFields.h"
 #include "SentryCrashReportVersion.h"
@@ -605,19 +604,6 @@ isRestrictedClass(const char *name)
     return false;
 }
 
-static void
-writeZombieIfPresent(
-    const SentryCrashReportWriter *const writer, const char *const key, const uintptr_t address)
-{
-#if SentryCrashCRASH_HAS_OBJC
-    const void *object = (const void *)address;
-    const char *zombieClassName = sentrycrashzombie_className(object);
-    if (zombieClassName != NULL) {
-        writer->addStringElement(writer, key, zombieClassName);
-    }
-#endif
-}
-
 static bool
 writeObjCObject(const SentryCrashReportWriter *const writer, const uintptr_t address, int *limit)
 {
@@ -700,7 +686,6 @@ writeMemoryContents(const SentryCrashReportWriter *const writer, const char *con
     writer->beginObject(writer, key);
     {
         writer->addUIntegerElement(writer, SentryCrashField_Address, address);
-        writeZombieIfPresent(writer, SentryCrashField_LastDeallocObject, address);
         if (!writeObjCObject(writer, address, limit)) {
             if (object == NULL) {
                 writer->addStringElement(
@@ -744,10 +729,6 @@ isNotableAddress(const uintptr_t address)
     const void *object = (const void *)address;
 
 #if SentryCrashCRASH_HAS_OBJC
-    if (sentrycrashzombie_className(object) != NULL) {
-        return true;
-    }
-
     if (sentrycrashobjc_objectType(object) != SentryCrashObjCTypeUnknown) {
         return true;
     }
@@ -1352,7 +1333,6 @@ writeError(const SentryCrashReportWriter *const writer, const char *const key,
 
         case SentryCrashMonitorTypeSystem:
         case SentryCrashMonitorTypeApplicationState:
-        case SentryCrashMonitorTypeZombie:
             SentryCrashLOG_ERROR(
                 "Crash monitor type 0x%x shouldn't be able to cause events!", crash->crashType);
             break;