+++ Sources/SentryCrash/Recording/Tools/SentryCrashMachineContext.c
@@ -147,6 +147,14 @@ void
 sentrycrashmc_suspendEnvironment(
     thread_act_array_t *suspendedThreads, mach_msg_type_number_t *numSuspendedThreads)
 {
+    sentrycrashmc_suspendEnvironment_upToMaxSupportedThreads(
+        suspendedThreads, numSuspendedThreads, UINT32_MAX);
+}
+
+void
+sentrycrashmc_suspendEnvironment_upToMaxSupportedThreads(thread_act_array_t *suspendedThreads,
+    mach_msg_type_number_t *numSuspendedThreads, mach_msg_type_number_t maxSupportedThreads)
+{
 #if SentryCrashCRASH_HAS_THREADS_API
     SentryCrashLOG_DEBUG("Suspending environment.");
     kern_return_t kr;
@@ -155,6 +163,12 @@ sentrycrashmc_suspendEnvironment(
 
     if ((kr = task_threads(thisTask, suspendedThreads, numSuspendedThreads)) != KERN_SUCCESS) {
         SentryCrashLOG_ERROR("task_threads: %s", mach_error_string(kr));
+        return;
+    }
+
+    if (*numSuspendedThreads > maxSupportedThreads) {
+        *numSuspendedThreads = 0;
+        SentryCrashLOG_DEBUG("Too many threads to suspend. Aborting operation.");
         return;
     }